requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
deprecation==2.1.0
grpcio==1.54.0
idna==3.4
numpy==1.24.3
packaging==23.1
protobuf==3.20.3
pydantic==1.10.7
pyTelegramBotAPI==4.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.29.0
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b54
typing_extensions==4.5.0
urllib3==1.26.15
